Minor update to Custom Stages Modeling
Anyone interested in making new custom stages presumably is aware there are new custom stages available for v3.  The language of the intro paragraph made it sound like this was impossible, and think might have created confusion, as likely an assumption that  this wiki page was obsolete and the template available somewhere. Hopefully this clears that up.